Fix Vars.
